% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.calculate.deviance.R
\name{calculate.gdm.deviance}
\alias{calculate.gdm.deviance}
\title{Calculate GDM deviance for observed & predicted dissimilarities}
\usage{
calculate.gdm.deviance(predDiss, obsDiss)
}
\arguments{
\item{predDiss}{(float) A vector of predicted dissimilarity values, of same
length as obsDiss.}

\item{obsDiss}{(float) A vector of observed dissimilarity values, of same
length as predDiss.}
}
\value{
A single value (float) being the deviance.
}
\description{
Calculate GDM deviance for observed & predicted dissimilarities.
Can be used for assessing cross-validation data. Translated from the c++
function CalcGDMDevianceDouble() in the file NNLS_Double.cpp from the
GDM R package.
}
