% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.partition.deviance.R
\name{gdm.partition.deviance}
\alias{gdm.partition.deviance}
\title{Perform Deviance Partitioning of a Fitted GDM}
\usage{
gdm.partition.deviance(sitePairTable, varSets=list(), partSpace=T)
}
\arguments{
\item{sitePairTable}{A correctly formatted site-pair table from
\code{\link[gdm]{formatsitepair}}.}

\item{varSets}{A list in which each element is a vector of variable names
across which deviance partitioning is to be performed, excluding
geographic distance (which is set by the partSpace argument). Variable names
must match those used to build the site-pair table. See example.}

\item{partSpace}{Whether or not to perform the partitioning using
geographic space. Default=TRUE.}
}
\value{
A dataframe summarizing partitioning results.
}
\description{
Partitions deviance explained into different components
(e.g., soils and climate), plus (optionally) space.
}
\examples{
# set up site-pair table
load(system.file("./data/southwest.RData", package="gdm"))
sppData <- southwest[c(1,2,13,14)]
envTab <- southwest[c(2:ncol(southwest))]
sitePairTab <- formatsitepair(sppData, 2, XColumn="Long", YColumn="Lat",
sppColumn="species", siteColumn="site", predData=envTab)

# Make list of variable sets for partitioning
varSet <- vector("list", 3)
varSet[[1]] <- c("awcA", "phTotal", "sandA", "shcA", "solumDepth")
varSet[[2]] <- c("bio5", "bio6")#, "bio15", "bio18", "bio19")
varSet[[3]] <- c("bio15", "bio18", "bio19")

# variable name sets
names(varSet) <- c("soil", "climate")
# names(varSet) <- c("soil", "temp", "precip")

# partition soils, climate, and space
scgPart <- gdm.partition.deviance(sitePairTab, varSet, partSpace=T)

# partition soils and climate only
scPart <- gdm.partition.deviance(sitePairTab, varSet, partSpace=F)

}
\author{
Matt Fitzpatrick and Karel Mokany
}
