% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdmDissim}
\alias{gdmDissim}
\title{An example biological dissimilarity matrix}
\format{
A data frame with 94 rows and 94 columns:
}
\usage{
gdmDissim
}
\description{
Pairwise Bray-Curtis dissimilarity calculated using the species
occurrence data from the \code{\link[gdm]{southwest}} data set.
}
\keyword{datasets}
