% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.permutate.sitepairtable.R
\name{permutateSitePair}
\alias{permutateSitePair}
\title{Permutate Site-Pair Table Rows}
\usage{
permutateSitePair(spTab, siteVarTab, indexTab, vNames)
}
\arguments{
\item{spTab}{A site-pair table.}

\item{siteVarTab}{A site x variable table.}

\item{indexTab}{A table of index values for the site-pair table.}

\item{vNames}{Vector of variable names in both the site-pair and
site x variable tables.}
}
\value{
A new site-pair table with rows randomized.
}
\description{
A function which randomizes the rows of a given site-pair table. This function
is called from the \code{\link[gdm]{gdm.varImp}} function and not needed by the user.
be called directly by the user.
}
\note{
This function is called from the \code{\link[gdm]{gdm.varImp}} function
and not needed by the user.
}
\seealso{
\code{\link[gdm]{gdm.varImp}}
}
\keyword{gdm}
\keyword{internal}
