% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.permutate.variable.R
\name{permutateVarSitePair}
\alias{permutateVarSitePair}
\title{Permutate the Values of a Site-Pair Table Variable}
\usage{
permutateVarSitePair(spTab, siteVarTab, indexTab, vName)
}
\arguments{
\item{spTab}{A given site-pair table.}

\item{siteVarTab}{A given site x variable table.}

\item{indexTab}{A table of index values for the given site-pair table.}

\item{vName}{Vector of variable names in both the site-pair and site-by-variable
tables.}
}
\value{
A site-pair table, with variable values randomized from the original
site-pair table.
}
\description{
A function which randomizes the values of the given variables
within a site-pair table. This function is called from the
\code{\link[gdm]{gdm.varImp}} function and should not need to be called directly
by the user.
}
\note{
This function is called from the \code{\link[gdm]{gdm.varImp}} function and
the user should not need to access directly.
}
\seealso{
\code{\link[gdm]{gdm.varImp}}
}
\keyword{gdm}
\keyword{internal}
