% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{southwest}
\alias{southwest}
\title{Species and environmental data from southwestern Australia.}
\format{
A data frame with 29364 rows and 14 variables:
\describe{
\item{species}{species name}
\item{site}{site name}
\item{awcA}{weight of the diamond, in carats}
\item{phTotal}{weight of the diamond, in carats}
\item{sandA}{weight of the diamond, in carats}
\item{shcA}{weight of the diamond, in carats}
\item{solumDepth}{weight of the diamond, in carats}
\item{bio5}{weight of the diamond, in carats}
\item{bio6}{weight of the diamond, in carats}
\item{bio15}{weight of the diamond, in carats}
\item{bio18}{weight of the diamond, in carats}
\item{bio19}{weight of the diamond, in carats}
\item{Lat}{weight of the diamond, in carats}
\item{Long}{weight of the diamond, in carats}
}
}
\usage{
southwest
}
\description{
A data set containing species occurrence and associated environmental
data at 94 sites in southwestern Australia.
}
\keyword{datasets}
