\name{stackedVars}
\alias{stackedVars}
\docType{data}
\title{
Worldclim climate variables used to fit GDM in examples
}
\description{
A raster stack of five Worldclim gridded climate variables used as environmental data in the GDM examples. The data were originally downloaded from Worldclim using the getData() function in the raster package and clipped to the area encompassing the example data points.
}
\format{
A raster stack of five layers.
}
\details{
The specific variables included in this data are: bio5, bio6, bio15, bio18, and bio19. The clipped area is southwestern Australia.
}
\source{
http://www.worldclim.org/
}
\references{
Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. Very high resolution interpolated climate surfaces for global land areas. International Journal of Climatology 25: 1965-1978.
}
\keyword{datasets}
