\name{gdmDissim}
\alias{gdmDissim}
\docType{data}
\title{
A sample biological dissimilarity matrix for use in GDM examples
}
\description{
A full dissimilarity matrix. 
}
\format{
A data frame with 180 rows and 180 columns.
}
\source{
Created using the example species data from southwest Australia.
}
\keyword{datasets}
